- Changed datatypes of Schedule and Employee ID's to long - added DemoHelper that wipes and fills the DB on each run, so we have an established development state. - Added HopperDao for general queries against the DB, basically added just so I can use a raw query to reset the Schedule and Employee tables on startup.

+++ app/src/main/java/com/example/f23hopper/data/schedule/Schedule.kt
@@ -19,8 +19,8 @@ import java.sql.Date
 )
 data class Schedule(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0,
     val date: Date,
-    val employeeId: Int,
+    val employeeId: Long,
     val shiftTypeId: Int
 )
